redbot-be/requirements.txt
Django==5.1.8
djangorestframework==3.15.2
drf-spectacular==0.27.2
requests==2.32.3
python-dotenv==1.0.1
